Resolve Bean overriding issues
 - Split the repositories into two, sub-packages (map and jpa) to prevent double scanning while canning for Map and Jpa repositories
 - Ensure that the Skipper auto configuration is started before the EorrorMvcAutoconfiguration
 - Temporarily disable Skipper OAuth configuration in favor of SCDF like security configuration.

 Resolves #756
