+++ spring-cloud-skipper-autoconfigure/src/main/java/org/springframework/cloud/skipper/server/autoconfigure/SkipperServerAutoConfiguration.java
@@ -15,8 +15,10 @@
  */
 package org.springframework.cloud.skipper.server.autoconfigure;
 
+import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.boot.autoconfigure.web.servlet.error.ErrorMvcAutoConfiguration;
 import org.springframework.cloud.skipper.server.config.EnableSkipperServerConfiguration;
 import org.springframework.cloud.skipper.server.config.SkipperServerConfiguration;
 import org.springframework.cloud.skipper.server.config.SkipperServerPlatformConfiguration;
@@ -31,6 +33,7 @@ import org.springframework.context.annotation.Import;
  */
 @Configuration
 @ConditionalOnBean(EnableSkipperServerConfiguration.Marker.class)
+@AutoConfigureBefore(ErrorMvcAutoConfiguration.class)
 @Import({SkipperServerConfiguration.class, SkipperServerPlatformConfiguration.class,
 		SpringDataRestConfiguration.class})
 @ConditionalOnProperty(prefix = "spring.cloud.skipper.server", name = "enabled", havingValue = "true", matchIfMissing = true)
