+++ spring-cloud-skipper-server-core/src/main/java/org/springframework/cloud/skipper/server/config/SkipperServerConfiguration.java
@@ -63,11 +63,11 @@ import org.springframework.cloud.skipper.server.deployer.strategies.UpgradeStrat
 import org.springframework.cloud.skipper.server.index.PackageMetadataResourceProcessor;
 import org.springframework.cloud.skipper.server.index.PackageSummaryResourceProcessor;
 import org.springframework.cloud.skipper.server.index.SkipperLinksResourceProcessor;
-import org.springframework.cloud.skipper.server.repository.AppDeployerDataRepository;
-import org.springframework.cloud.skipper.server.repository.DeployerRepository;
-import org.springframework.cloud.skipper.server.repository.PackageMetadataRepository;
-import org.springframework.cloud.skipper.server.repository.ReleaseRepository;
-import org.springframework.cloud.skipper.server.repository.RepositoryRepository;
+import org.springframework.cloud.skipper.server.repository.jpa.AppDeployerDataRepository;
+import org.springframework.cloud.skipper.server.repository.jpa.PackageMetadataRepository;
+import org.springframework.cloud.skipper.server.repository.jpa.ReleaseRepository;
+import org.springframework.cloud.skipper.server.repository.jpa.RepositoryRepository;
+import org.springframework.cloud.skipper.server.repository.map.DeployerRepository;
 import org.springframework.cloud.skipper.server.service.PackageMetadataService;
 import org.springframework.cloud.skipper.server.service.PackageService;
 import org.springframework.cloud.skipper.server.service.ReleaseReportService;
@@ -102,10 +102,9 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 @EnableConfigurationProperties({ SkipperServerProperties.class, VersionInfoProperties.class,
 		LocalPlatformProperties.class, /* CloudFoundryPlatformProperties.class, */ MavenConfigurationProperties.class,
 		HealthCheckProperties.class })
-@EntityScan({ "org.springframework.cloud.skipper.domain",
-		"org.springframework.cloud.skipper.server.domain" })
-@EnableMapRepositories(basePackages = "org.springframework.cloud.skipper.server.repository")
-@EnableJpaRepositories(basePackages = "org.springframework.cloud.skipper.server.repository")
+@EntityScan({ "org.springframework.cloud.skipper.domain", "org.springframework.cloud.skipper.server.domain" })
+@EnableMapRepositories(basePackages = "org.springframework.cloud.skipper.server.repository.map")
+@EnableJpaRepositories(basePackages = "org.springframework.cloud.skipper.server.repository.jpa")
 @EnableTransactionManagement
 @EnableAsync
 @Import({ StateMachinePersistConfiguration.class, StateMachineExecutorConfiguration.class,
