+++ spring-cloud-skipper-server-core/src/main/java/org/springframework/cloud/skipper/server/repository/jpa/AppDeployerDataRepositoryImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 the original author or authors.
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.cloud.skipper.server.repository;
+package org.springframework.cloud.skipper.server.repository.jpa;
 
 import java.util.List;
 import java.util.stream.Collectors;
