+++ spring-cloud-skipper-server-core/src/test/java/org/springframework/cloud/skipper/server/repository/RepositoryCreator.java
@@ -16,6 +16,7 @@
 package org.springframework.cloud.skipper.server.repository;
 
 import org.springframework.cloud.skipper.domain.Repository;
+import org.springframework.cloud.skipper.server.repository.jpa.RepositoryRepository;
 
 /**
  * @author Mark Pollack
